requirements.txt
Flask>=2.3.0
Flask-Cors>=4.0.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
